% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfboin-package.R
\docType{package}
\name{bfboin-package}
\alias{bfboin}
\alias{bfboin-package}
\title{bfboin: Operating Characteristics for the Bayesian Optimal Interval Design with Back Filling}
\description{
Calculate the operating characteristics of the Bayesian Optimal Interval with Back Filling Design for dose escalation in early-phase oncology trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openpharma.github.io/bfboin/}
}

}
\author{
\strong{Maintainer}: Dominic Magirr \email{dominic.magirr@novartis.com} [copyright holder]

Authors:
\itemize{
  \item Bairu Zhang \email{bairu.zhang@novartis.com}
}

}
\keyword{internal}
